pyproject.toml
[tool.poetry]
name = "oaisamples"
version = "0.1.0"
description = ""
authors = ["Michael Ikemann <michael@omitted.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^1.3.8"
tiktoken = "^0.5.2"
pydantic = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
